requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai==0.8.3
requests
loguru
